requirements.txt
asyncio==3.4.3
bitstring==3.1.5
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django>=2.1.7
django-redis==4.10.0
djangorestframework==3.9.1
drf-yasg==1.13.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2==2.7.7
pytz==2018.9
redis==3.2.0
requests==2.21.0
ruamel.yaml==0.15.88
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.18
pycryptodome>=3.7.3
SQLAlchemy>=1.3.1
aiopg>=0.16
SQLAlchemy-Utils>=0.33.11
gino>=0.8.2
jsonpickle==1.1
aioredis==1.2.0